LinkId üzerinden sorgulama taammlandı. #42
