+++ iParaClientService.Integration.Tests/iParaLinkPaymentCreateAdapterTests.cs
@@ -9,7 +9,7 @@ using iParaClientService.Service;
 namespace iParaClientService.Integration.Tests
 {
     [TestClass]
-    public class iParaLinkPaymentListAdapterTests
+    public class iParaLinkPaymentCreateAdapterTests
     {
         private string publicKey;
         private string privateKey;
@@ -29,7 +29,7 @@ namespace iParaClientService.Integration.Tests
 
         [TestMethod]
 
-        public void Can_List_Link_Payment_Valid()
+        public void Can_Create_Link_Payment_Valid()
         {
             string baseUrl = "https://api.ipara.com/";
             iParaConnectionMode mode = iParaConnectionMode.Test;
@@ -37,16 +37,25 @@ namespace iParaClientService.Integration.Tests
 
             using (var connection = new iParaClientConnection(baseUrl, publicKey, privateKey, mode, version))
             {
-                var model = new iParaLinkPaymentListRequest
+                var model = new iParaLinkPaymentCreateRequest
                 {
+                    ThreeD = false,
+                    ClientIp = "127.0.0.1",
+                    CommissionType = CommissionType.Dealer,
+                    Email = "kurtayyilmaz@gmail.com",
+                    SendEmail = true,
+                    Name = "Kurtay",
+                    Surname = "Yılmaz",
+                    TcCertificate = "18946604794",
+                    TaxNumber = "",
                     Echo = "",
                     Mode = "T",
-                    PageSize = 15,
-                    PageIndex = 1,
-                    Email = "kurtayyilmaz@gmail.com"
+                    ExpireDate = DateTime.Now.AddDays(1).ToString("yyyy-MM-dd hh:mm:ss"),
+                    Gsm = "5397143516",
                 };
+                model.SetAmount(10);
 
-                var adapter = new iParaLinkPaymentListAdapter(connection);
+                var adapter = new iParaLinkPaymentCreateAdapter(connection);
                 var result = adapter.Execute(model);
 
                 Trace.WriteLine(result.ErrorMessage);
@@ -55,8 +64,9 @@ namespace iParaClientService.Integration.Tests
                 Assert.IsNull(result.ErrorCode);
                 Assert.IsNull(result.ErrorMessage);
 
-                Assert.IsNotNull(result.LinkPaymentRecordList);
-                Assert.IsTrue(result.LinkPaymentRecordList.Count > 0);
+                Assert.AreEqual(model.Amount, result.Amount.Value);
+                Assert.IsNotNull(result.Link);
+                Assert.IsNotNull(result.LinkPaymentId);
                 Assert.IsNotNull(result.Result);
                 Assert.IsNotNull(result.ResponseMessage);
             }
@@ -64,7 +74,7 @@ namespace iParaClientService.Integration.Tests
 
         [TestMethod]
 
-        public void Can_List_Link_Payment_Invalid()
+        public void Can_Create_Link_Payment_Invalid()
         {
             string baseUrl = "https://api.ipara.com/";
             iParaConnectionMode mode = iParaConnectionMode.Test;
@@ -72,14 +82,24 @@ namespace iParaClientService.Integration.Tests
 
             using (var connection = new iParaClientConnection(baseUrl, publicKey, privateKey, mode, version))
             {
-                var model = new iParaLinkPaymentListRequest
+                var model = new iParaLinkPaymentCreateRequest
                 {
+                    ThreeD = true,
                     ClientIp = "127.0.0.1",
+                    CommissionType = CommissionType.Dealer,
+                    Email = "kurtayyilmaz@gmail.com",
+                    SendEmail = true,
+                    Name = "Kurtay",
+                    Surname = "Yılmaz",
+                    TcCertificate = "",
+                    TaxNumber = "",
                     Echo = "",
                     Mode = "T",
+                    ExpireDate = DateTime.Now.AddDays(1).ToString("yyyy-MM-dd hh:mm:ss"),
                     Gsm = "53143516",
                 };
-                var adapter = new iParaLinkPaymentListAdapter(connection);
+                model.SetAmount(10);
+                var adapter = new iParaLinkPaymentCreateAdapter(connection);
                 var result = adapter.Execute(model);
 
                 Trace.WriteLine(result.ErrorMessage);
@@ -87,8 +107,10 @@ namespace iParaClientService.Integration.Tests
                 Assert.IsFalse(result.IsValid);
                 Assert.IsNotNull(result.ErrorCode);
                 Assert.IsNotNull(result.ErrorMessage);
+                Assert.IsNull(result.Link);
+                Assert.IsNull(result.LinkPaymentId);
                 Assert.IsNotNull(result.Result);
-
+                
             }
         }
