+++ iParaClientService.Integration.Tests/iParaLinkPaymentListAdapterTests.cs
@@ -1,15 +1,18 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
 using System.Diagnostics;
+using System.Linq;
+using FluentAssertions;
 using iParaClientService.Adapter;
 using iParaClientService.Model;
 using iParaClientService.Model.Request;
 using iParaClientService.Service;
+using iParaClientService.Utils;
 
 namespace iParaClientService.Integration.Tests
 {
     [TestClass]
-    public class iParaLinkPaymentCreateAdapterTests
+    public class iParaLinkPaymentListAdapterTests
     {
         private string publicKey;
         private string privateKey;
@@ -29,7 +32,7 @@ namespace iParaClientService.Integration.Tests
 
         [TestMethod]
 
-        public void Can_Create_Link_Payment_Valid()
+        public void Can_List_Link_Payment_Valid()
         {
             string baseUrl = "https://api.ipara.com/";
             iParaConnectionMode mode = iParaConnectionMode.Test;
@@ -37,25 +40,16 @@ namespace iParaClientService.Integration.Tests
 
             using (var connection = new iParaClientConnection(baseUrl, publicKey, privateKey, mode, version))
             {
-                var model = new iParaLinkPaymentCreateRequest
+                var model = new iParaLinkPaymentListRequest
                 {
-                    ThreeD = false,
-                    ClientIp = "127.0.0.1",
-                    CommissionType = CommissionType.Dealer,
-                    Email = "kurtayyilmaz@gmail.com",
-                    SendEmail = true,
-                    Name = "Kurtay",
-                    Surname = "Yılmaz",
-                    TcCertificate = "18946604794",
-                    TaxNumber = "",
                     Echo = "",
                     Mode = "T",
-                    ExpireDate = DateTime.Now.AddDays(1).ToString("yyyy-MM-dd hh:mm:ss"),
-                    Gsm = "5397143516",
+                    PageSize = 15,
+                    PageIndex = 1,
+                    Email = "kurtayyilmaz@gmail.com"
                 };
-                model.SetAmount(10);
 
-                var adapter = new iParaLinkPaymentCreateAdapter(connection);
+                var adapter = new iParaLinkPaymentListAdapter(connection);
                 var result = adapter.Execute(model);
 
                 Trace.WriteLine(result.ErrorMessage);
@@ -64,9 +58,8 @@ namespace iParaClientService.Integration.Tests
                 Assert.IsNull(result.ErrorCode);
                 Assert.IsNull(result.ErrorMessage);
 
-                Assert.AreEqual(model.Amount, result.Amount.Value);
-                Assert.IsNotNull(result.Link);
-                Assert.IsNotNull(result.LinkPaymentId);
+                Assert.IsNotNull(result.LinkPaymentRecordList);
+                Assert.IsTrue(result.LinkPaymentRecordList.Count > 0);
                 Assert.IsNotNull(result.Result);
                 Assert.IsNotNull(result.ResponseMessage);
             }
@@ -74,7 +67,7 @@ namespace iParaClientService.Integration.Tests
 
         [TestMethod]
 
-        public void Can_Create_Link_Payment_Invalid()
+        public void Can_List_Link_Payment_Valid_ByState()
         {
             string baseUrl = "https://api.ipara.com/";
             iParaConnectionMode mode = iParaConnectionMode.Test;
@@ -82,24 +75,92 @@ namespace iParaClientService.Integration.Tests
 
             using (var connection = new iParaClientConnection(baseUrl, publicKey, privateKey, mode, version))
             {
-                var model = new iParaLinkPaymentCreateRequest
+                var model = new iParaLinkPaymentListRequest
                 {
-                    ThreeD = true,
-                    ClientIp = "127.0.0.1",
-                    CommissionType = CommissionType.Dealer,
-                    Email = "kurtayyilmaz@gmail.com",
-                    SendEmail = true,
-                    Name = "Kurtay",
-                    Surname = "Yılmaz",
-                    TcCertificate = "",
-                    TaxNumber = "",
                     Echo = "",
                     Mode = "T",
-                    ExpireDate = DateTime.Now.AddDays(1).ToString("yyyy-MM-dd hh:mm:ss"),
+                    PageSize = 15,
+                    PageIndex = 1,
+                    Email = "kurtayyilmaz@gmail.com",
+                    LinkState = LinkState.LinkPaymentDone
+                };
+
+                var adapter = new iParaLinkPaymentListAdapter(connection);
+                var result = adapter.Execute(model);
+
+                Trace.WriteLine(result.ErrorMessage);
+
+                Assert.IsTrue(result.IsValid);
+                Assert.IsNull(result.ErrorCode);
+                Assert.IsNull(result.ErrorMessage);
+
+                Assert.IsNotNull(result.LinkPaymentRecordList);
+                Assert.IsTrue(result.LinkPaymentRecordList.Count > 0);
+                Assert.IsNotNull(result.Result);
+                Assert.IsNotNull(result.ResponseMessage);
+            }
+        }
+        [TestMethod]
+        [DataRow("002R/OgkSDc2BpGXb4cr79R5Q==", "100")]
+        public void Can_List_Link_Payment_Valid_ByLinkId(string linkId, string amount)
+        {
+            string baseUrl = "https://api.ipara.com/";
+            iParaConnectionMode mode = iParaConnectionMode.Test;
+            string version = "1.0";
+
+            using (var connection = new iParaClientConnection(baseUrl, publicKey, privateKey, mode, version))
+            {
+                var model = new iParaLinkPaymentListRequest
+                {
+                    Echo = "",
+                    Mode = "T",
+                    PageSize = 15,
+                    PageIndex = 1,
+                    Email = "kurtayyilmaz@gmail.com",
+                    LinkId = linkId,
+                    LinkState = LinkState.LinkPaymentDone
+                };
+
+                var adapter = new iParaLinkPaymentListAdapter(connection);
+                var result = adapter.Execute(model);
+
+                Trace.WriteLine(result.ErrorMessage);
+
+                Assert.IsTrue(result.IsValid);
+                Assert.IsNull(result.ErrorCode);
+                Assert.IsNull(result.ErrorMessage);
+
+                Assert.IsNotNull(result.LinkPaymentRecordList);
+                Assert.IsTrue(result.LinkPaymentRecordList.Count == 1);
+
+                var payment = result.LinkPaymentRecordList.Single();
+                payment.Amount.Should().Be(Convert.ToInt16(amount));
+                payment.LinkId.Should().Be(linkId);
+                payment.LinkState.Should().Be(LinkState.LinkPaymentDone);
+
+                Assert.IsNotNull(result.Result);
+                Assert.IsNotNull(result.ResponseMessage);
+            }
+        }
+
+        [TestMethod]
+
+        public void Can_List_Link_Payment_Invalid()
+        {
+            string baseUrl = "https://api.ipara.com/";
+            iParaConnectionMode mode = iParaConnectionMode.Test;
+            string version = "1.0";
+
+            using (var connection = new iParaClientConnection(baseUrl, publicKey, privateKey, mode, version))
+            {
+                var model = new iParaLinkPaymentListRequest
+                {
+                    ClientIp = "127.0.0.1",
+                    Echo = "",
+                    Mode = "T",
                     Gsm = "53143516",
                 };
-                model.SetAmount(10);
-                var adapter = new iParaLinkPaymentCreateAdapter(connection);
+                var adapter = new iParaLinkPaymentListAdapter(connection);
                 var result = adapter.Execute(model);
 
                 Trace.WriteLine(result.ErrorMessage);
@@ -107,10 +168,8 @@ namespace iParaClientService.Integration.Tests
                 Assert.IsFalse(result.IsValid);
                 Assert.IsNotNull(result.ErrorCode);
                 Assert.IsNotNull(result.ErrorMessage);
-                Assert.IsNull(result.Link);
-                Assert.IsNull(result.LinkPaymentId);
                 Assert.IsNotNull(result.Result);
-                
+
             }
         }
